[feat] Adds the native functions "type", "str", and "num".
The "type" function returns a string indicating the type of the argument. The "str" function converts an argument to a string, while the "num" function converts an argument to a number.

+++ src/lib/string/yapl_string.c
@@ -12,10 +12,10 @@
 /**
  * Hashes an input string using FNV-1a hash function.
  */
-uint32_t hashString(const char *key, int length) {
+uint32_t hashString(const char *key, uint64_t length) {
     uint32_t hash = 2166136261u;
 
-    for (int i = 0; i < length; i++) {
+    for (uint64_t i = 0; i < length; i++) {
         hash ^= key[i];
         hash *= 16777619;
     }
@@ -37,7 +37,7 @@ ObjString *makeString(int length) {
  * Copies and allocates a given string to the heap. This way, every ObjString reliably owns its
  * character array and can free it.
  */
-ObjString *copyString(const char *chars, int length) {
+ObjString *copyString(const char *chars, uint64_t length) {
     uint32_t hash = hashString(chars, length);
     ObjString *interned = tableFindStr(&vm.strings, chars, length, hash); /* Checks if interned */
     if (interned != NULL) return interned;
